Fix clang-tidy `HeaderFilterRegex`.
In general LLVM's regexp engine does not seem to support negative
matches at all. We previously tried to set this up, but this also
filtered many headers we would have liked to lint.

This patch sets a regexp under the assumptions that our paths do not
contain numbers. With that many more diagnostics are triggered which we
also fix.

+++ spicy/toolchain/include/compiler/detail/codegen/productions/type-literal.h
@@ -26,7 +26,7 @@ public:
     bool nullable() const { return false; }
     bool eodOk() const { return nullable(); }
     bool atomic() const { return true; }
-    int64_t tokenID() const { return production::tokenID(hilti::util::fmt("%s", _type)); }
+    int64_t tokenID() const { return static_cast<int64_t>(production::tokenID(hilti::util::fmt("%s", _type))); }
     std::string render() const { return hilti::util::fmt("%s", _type); }
 
 private:
